requirements.txt
pyyaml==6.0.2
fastapi==0.115.0
uvicorn==0.30.6
